feat: Only FIP editor show create button
